+++ ecomarket/src/main/java/com/ecomarket/ecomarket/repository/ProductoRepository.java
@@ -10,10 +10,10 @@ import java.util.Optional;
 @Repository
 public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
-    Optional<Producto> findByIdProducto(Integer idProducto);
+    Optional<Producto> findByIdProducto(Long idProducto);
     Optional<Producto> findByNombre(String nombre);
     List<Producto> findByPrecio(double precio);
-    List<Producto> findByStock(Integer stock);
+    List<Producto> findByStock(Long stock);
     List<Producto> findByCategoria(String categoria);
     List<Producto> findByDescuento(double descuento);
